Fixed post requests, added session verification and logout
Post requests now are sent a max of 3 times if they fail.

The user's cookies are checked on each page to ensure that they are still valid, and that the user has the proper authorization to be on the page.

The logout button is now functional and removes all of a user's cookies upon being pressed.